Move TranslateFn to shared @/types/i18n module
