+++ frontend/src/components/projects/KanbanColumn.tsx
@@ -10,7 +10,8 @@ import { Button } from "@/components/ui/button";
 import { Markdown } from "@/components/Markdown";
 import type { ProjectTaskStatus, Task, TaskPriority } from "@/types/api";
 import { truncateText } from "@/lib/text";
-import { summarizeRecurrence, type TranslateFn } from "@/lib/recurrence";
+import { summarizeRecurrence } from "@/lib/recurrence";
+import type { TranslateFn } from "@/types/i18n";
 import { TaskAssigneeList } from "@/components/projects/TaskAssigneeList";
 import { cn } from "@/lib/utils";
 import { TaskChecklistProgress } from "@/components/tasks/TaskChecklistProgress";
