+++ frontend/src/types/i18n.ts
@@ -1,0 +1,2 @@
+/** Loose translation function signature compatible with any i18next namespace. */
+export type TranslateFn = (key: string, options?: Record<string, unknown>) => string;
